Move Struct test helper class to a helpers module so that we can use it elsewhere more easily
